python-twitter-importer/requirements.txt
nltk
sentiment_classifier
pymongo
tweepy
python-dateutil

